Add abstracttext as submodule
Clone with `--recurse-submodules`, or in an existing clone run `git
submodule --init --recursive`, to clone the submodule as well.
Evaluating a reference node now tries to load JSON files from the
submodule. That only works if the current working directory is the
GraalEneyj checkout, but hey, it’s better than the hard-coded path to
somewhere in my home directory. I’ll figure the rest out later.