Add ZReference runtime object
Instead of immediately evaluating a reference node, make it return a
runtime object which can be evaluated on-demand; for interop, this is
exposed as a function with no arguments.

ZReference is not added to ZTypes because, judging from SimpleLanguage,
it doesn’t look like all runtime types have to be registered there. (I’m
not sure what the rules for this are, but for now, not registering
ZReference keeps things simpler.)