Add rudimentary function call support
Taken together, this is enough to evaluate the following function call:

    { "Z1K1": "Z7", "Z7K1": "Z36", "K1": "Z28" }

Which in the eneyj CLI (assuming the English language) would read as:

    value(project_name)

and evaluates to "eneyj".

The parser learns to parse functions (with their implementations) and
function calls, though for now the only supported implementation type is
the builtin, and the only known builtin is “value”.

Function and implementation runtime objects are added. An implementation
wraps a single RootCallTarget; a function executes by delegating to one
of its implementations, which can change at runtime.

An abstract class for builtin nodes is added, and the value builtin is
the first subclass. Currently, it hard-codes the value for the Z28
(project_name) reference, and does nothing special for all other values.

A test for this function call is added as well.