Don’t hard-code Z28 project_name
Instead of hard-coding the value result for that one reference, evaluate
the reference, and if it’s a string, return the string value. Still no
support for anything else.

I’m also not sure if any of this will stay. Maybe object literals of
type Z6 will be reflected as a special runtime type, so they can act as
strings for interop (but without losing the extra properties). Also,
maybe references won’t be runtime objects anymore, but will actually be
evaluated eagerly. I don’t know. But still, this is a step forwards.