Support Global and Local Keys
Global Keys are ones that start with a ZID and correspond to named
arguments, whereas Local Keys are ones that start directly with the K
and correspond to positional arguments. Parse and support both. (In
theory, this should work for function calls where the function isn’t a
reference, but I don’t think I can test that yet. Also, it might not do
the right thing for function calls where the function is a reference,
but not a global one.)