Add support for “if” structure
Conceptually, this is a function call, but we make it a special node for
lazy evaluation (builtin functions are currently called with
already-evaluated arguments) and condition profiling.

The testIfValueTrue() test corresponds to this expression:

    if(value(true), "then", "else")

AbstractText’s eneyj evaluates this to "else" rather than our "then"; I
believe this to be a bug there, reported as google/abstracttext#21.