Support code implementations
So far, we try to support two languages, JavaScript and Python – these
are the only ones used in upstream eneyj, though in principle we could
support others too. That said, GraalPython doesn’t seem to support parse
requests with argument names yet, so for now that doesn’t work and the
test is ignored.

The ZContext gains an Env and a method which can send a parse request to
that Env. Code implementation nodes are a new node type that uses this
method to create the call target, using hard-coded argument names for
now, and the parser learns to parse these nodes. The launcher and ZTest
are both updated to configure their context in such a way as to allow
polyglot access, because otherwise we can’t access those languages
ourselves. The tests use a custom version of Z53 (the nand function) for
now, since we can’t parse the real one yet (unsupported implementation
type Z7). To run the tests, GraalJS is added as a Maven dependency (the
JS tests still skip themselves if JS isn’t installed, but we don’t
expect this to happen); GraalPython isn’t available via Maven yet and
instead added to the class path from the JAVA_HOME, along with Sulong
(the LLVM language, which it requires).