Support automatic switching between implementations
Implementations can now throw an UnusableImplementationException to
signal that they’re unusable, and the ZFunction catches this and
switches to a different implementation, as long as usable ones remain.
We use this for code in unsupported languages, unrecognized builtins,
and unsupported implementation types (such as Z7 functioncall
implementations).

In theory, this should allow us to switch the CodeJsTest from our custom
nandWithOnlyJsImplementation() to the real Z53 function – previously
this threw an exception at parse time (unsupported implementation type
Z7), now it should throw an UnusableImplementationException and
automatically switch to the (usable) code implementation. However, in
practice this doesn’t work due to an unrelated issue, which is that code
loaded via a ZReference can’t be executed yet – it complains that the
node isn’t adopted. (I had also run into this issue previously when
trying to implement the Z7 implementation type – none of that experiment
has made it into Git history yet.) I suspect that the way we currently
execute references is deeply wrong, and only worked by accident so far
(because we called no method that required properly adopted nodes) – in
reality, reference execution should probably look more like code
implementation execution, instead of parsing a node out of thin air and
directly calling its execute() method. But I’ll fix that separately.