Add head (Z64) builtin
The builtin reads the first element of any list using the interop
protocol; maybe I’ll later add specializations for the specific ZList
type (I’m not sure if that’s even necessary or if the interop protocol
is cheap anyways).

This also includes a first stab at exception handling: the
ZErrorException type implements TruffleException and is *supposed* to
wrap an object like Z441 list_is_nil, but I haven’t fully implemented
that yet (I’m not sure how best to evaluate that object inside the
exception handler), so for now the inner error is null.