Implement hasLanguage(), getLanguage(), toDisplayString()
There was a deprecation warning for ZLanguage implementing the
isObjectOfLangnuage() method. The notice says to implement
getLanguageView() instead, but if I understand correctly, we don’t
actually need to do that, as long as all of our runtime values export
the right interop messages: hasLanguage(), getLanguage(), and
toDisplayString(). (Only hasLanguage() and getLanguage() are really
required for this, I assume, but they say that if you implement those,
you must also implement toDisplayString().)

The toDisplayString() implementations are mostly straightforward; the
more involved ones (list and object) are marked as truffle boundaries,
because I don’t think it would be useful to compile or inline them, they
shouldn’t be used during regular program execution.