Pass function ID into functions and implementations
This allows us to get rid of some absurd hard-coding in
ZImplementationCodeNode – instead, the node truly knows the argument
names of the function it implements, and passes those to the underlying
language runtime.

Adds a second set of JS code tests, using the negate function in
addition to the existing nand tests. (Still needs to be a custom
function with only a JS implementation, rather than the real Z56,
though.)