Evaluate references more properly
The old method was a bit ugly, and part of that was that we were
evaluating orphaned nodes which we had just parsed out of thin air.
Instead, loading a reference should work more like evaluating a code
implementation: ask Truffle to turn a Source into a CallTarget and then
call that.

This requires us to set the language home, and the source files we load
need to be below the language home (and we’re not allowed to hide that
behind symlinks either), so we need to swap the abstracttext submodule
and symlink so that the real abstracttext data is properly below
language/, which is the language home when running the unit tests (see
the -D in the language/ pom.xml).