Add Z33/same builtin
We implement it using equals(), which we implement in our runtime
objects. The builtin is also meant to call Z36/value on its arguments,
which we do not in the ZSameBuiltin node, but by building the AST in
such a way that the builtin’s arguments are always wrapped in a
ZValueBuiltin node. Astonishingly, this works! I really like that :)