Cache reference evaluation results
If a ZEvaluateReferenceNode is only ever called with one reference,
cache its result instead of having to look it up in the context. This
should optimize function calls (most of which call a constant ZFunction
– the call target of the function varies, but not the function itself!)
with minimal cost for the value builtin, since this specialization will
be replaced as soon as it sees a second, different reference. (In the
future, we could benchmark if higher limits result in better
performance, but to me a limit of 1 feels like the most useful.)