Register Z31/if as a builtin
We parse function calls to Z31/if specially, to enable lazy evaluation
of the consequent and alternative. (Originally, this was also motivated
by separate condition profiling, but I suspect this no longer matters so
much now that we use a fresh profile for each copy of the node.)
However, it’s possibly to have calls to Z31/if that the parser cannot
recognize, and we should still be able to execute those, so register the
ZIfNode also as a builtin implementation. (In this case, the consequent
and alternative will be both evaluated, but that cannot be helped, I
think; since functions are pure and have no side-effects, this only
affects performance, not correctness.)