Implement ZReference.execute() as specializations
Change the evaluate() callers to use an interop library to execute the
reference, and turn execute() into an abstract class with multiple
specializations. I hoped to directly return a cached value from the
reference, similar to ZEvaluateReferenceNode.doReferenceCached(), but
that didn’t work out and failed with errors I couldn’t figure out. This
leaves us with just two specializations: the “arguments.length > 0" case
and the “generic” case for everything else.