Use @Cached for ZReference execution
I figured out how to make it work after all – allowUncached = true was
the missing bit. (I had already noticed that this made the error go
away, but need to look at the generated source code before convincing
myself that this really is the right thing to do.)

I’m not sure if the extra guard for the cachedId is necessary. Removing
it doesn’t make any tests fail, but looking at the generated source
code, I’m not convinced removing it is safe. Add a reminder to
reconsider removing it in the future, when we have a lot more tests that
could inspire confidence in it really being unnecessary.