Parse function calls more liberally
If the function being called is a reference to another function, then
the argument names might be global keys, but not use the called function
as a prefix. For instance, the eneyj Z56/negate calls Z104/if_boolean
but with the keys Z31K1, Z31K2, Z31K3 (rather than Z104K1 etc.). To
parse this without knowing the relationship between Z104 and Z31, we
collect all keys (apart from Z1K* and Z7K* – Z1K1 is Z7/functioncall,
and Z7K1 is the function being called), assert they all use the same
prefix (be that the same as the called function, a different one, or the
empty string for local keys / positional arguments) and are contiguous,
and then use those for the arguments. For more discussion on this, see
also google/abstracttext#28 [1].

[1]: https://github.com/google/abstracttext/issues/28