Add character type and Z60/character_to_string builtin
A Z60/character is represented by a Java int, because a Java char is not
large enough to hold any Unicode code point. (If we ever want to add
primitive integers, we can still use long for that.) Teach the parser to
recognize {"Z1K1": "Z60", "Z60K1": "c"} literals as character literals
which evaluate to ints, and define the builtin to turn a character into
a string.

The Z61/character_to_string function actually defines the builtin as the
second implementation; the first implementation is based on Z37/reify
and Z106/find_second_by_first, but that implementation doesn’t work yet
(we haven’t implemented the Z37/reify builtin yet). Change ZFunction to
throw an UnusableImplementationException when all of its implementations
are exhausted, to signal to any *outer* function (in this case,
Z61/character_to_string) that it should proceed with the next
implementation (here, the builtin).

(I suppose this also means that once we do support Z37/reify, and the
first implementiaton works, the unit test added here would actually no
longer test the builtin implementation. Hopefully I’ll remember to
update the test then.)

The value builtin is also taught to turn a zobject of type Z60 into a
plain character (int), but there’s not really a way to test this yet –
there’s no way to *create* such a zobject at the moment. Eventually, I
should update the parser to parse Z60 literals with more keys than Z1K1
and Z60K1 as object literals, and then adjust other places to treat such
objects just like characters, but that’s more effort than I want to
spend just now.