Add ZCharacter runtime type
This should be used to box a character (int) instead of Integer. I’m not
sure how the various pieces here fit together yet, but it seems to work
more or less, at least.

At some point we can probably think about an instance cache for common
characters, but not right now.