Restore and update old ZTypes code
This code was commented out (originally coming from SimpleLanguage) back
when I still thought Z13/nil should be the “null” value for interop, but
in fact that role should be filled by Z23/nothing. This means that we
can use the code again, though it’s not actually used for anything at
the moment, since we have no @Specialization for ZNothing.