Add first version of Z37/reify builtin
It’s not complete yet, but it works on some simple inputs, at least.
I’ll continue working on it later.

The character_to_string tests are adjusted to use a custom version of
the function they’re testing, with only the builtin implementation,
because the real function has an alternative implementation that so far
was unusable (because Z37/reify was not implemented) but now is no
longer marked unusable; it still doesn’t work (because it ends up using
Z106/find_second_by_first, which is effectively implemented using
itself), but the error is now a stack overflow, so the non-builtin
implementation is never marked unusable and the function never falls
back to the builtin. But this makes sense anyways – the tests are
supposed to test specifically the builtin, not any other implementation
that might exist.