Fix ZValueBuiltin for ZCharacter inputs
For ZCharacter values, the builtin actually did nothing. Fix that and
write proper tests for it.

Also, add a missing cast in getValue(), though I suspect it’s actually
unnecessary and dead code, because there shouldn’t be any way to create
a ZObject of type character in the first place. But I’ll clean that up
later.