Add ZString runtime object
This is used to represent strings with extra members, instead of just
throwing those members away at parse time. The code is pretty similar to
the existing ZCharacter and ZCharacterLiteralNode.

Currently, a ZString must have at least one extra member, and a string
without members must be represented by a plain String. I’m not yet sure
if this is actually the best thing, or if we should in some way wrap
Strings in ZStrings to expose the default members, Z1K1 and Z6K1.