Don’t turn references to strings in Z37/reify
I did this for consistency with eneyj, but I don’t think it’s a good
idea, nor should it be necessary. References can stay references – this
will make it much easier for Z38/abstract to recognize them.

(Note that tests already used toString() where previously they could
have used asString(), because the values used to be proper strings, but
now they’re references and toString() needs to be used to get the
display string.)