Support Python implementations
GraalPython 20.2.0 supports parse requests with argument names, so we
can now support Python implementations. The test code needs to be
updated a bit (in the time since it was written, the ID and parameter
list of a function have become mandatory), but mostly it still works.

Annoyingly, this makes the tests rather slower – from some 2.5 seconds
to almost 12 seconds, on my system. Most of that is in CodePythonTest
(around 8 seconds alone), but FunctioncallTest also gets slower,
presumably because some Python implementations now become usable that
used to fall back to JS or function call implementations. Maybe that can
be improved later, but for now I can live with it. (If it gets too
annoying for local development, I can always uninstall GraalPython, I
suppose, and let Travis CI warn me if these tests break.)