Introduce assertZReference
Previously, to assert that a value was a reference for the correct ID,
we compared its .toString() result. But this succeeds for both
references and strings, so it doesn’t capture the full intent of the
test. Introduce a dedicated assertion for references; for now, it only
has one additional check (that the value isn’t a string), but I’d like
to improve it later (we probably need to implement the meta-object
interop messages for this).

This only replaces some uses of .toString(), because apparently some
places that are *supposed* to see references actually get strings. I’ll
fix that separately.