Add tests for same(value, abstract(reify(value)))
These tests are fairly simple to write, since the Z33/same builtin does
all the “work” of comparing the expected and actual value for us. The
downside is that if the test fails it will be with practically no useful
information – the builtin just returns false or true – so the
investigating developer will need to modify the test code to get more
information about why it’s failing.