Add test for recursive abstract behavior
Produce a slightly simplified version of a reified project_name, and
abstract it back into an object. Check that it looks as it should. (We
can’t properly handle a real reified project_name yet because
multilingual text involves a list, and I do not understand upstream
eneyj’s behavior for reifying lists at all.)