Support reifying lists
We don’t actually touch the reify builtin at all, we just implement the
needed interop messages on ZList. Some helpers are added to the tests to
assert reified lists relatively easily, similar to the helper we already
had for pairs.

This doesn’t fully match upstream eneyj, for two reasons:

1. We ignore the Z13.json file and hard-code the contents of the reified
   Z13/nil. It has an ID but no labels.
2. Upstream eneyj recursively reifies the Z441/list_is_nil error you get
   when you try to access the Z10K1/head or Z10K2/tail of Z13/nil. We
   just emit a reference to the error.

But I think it’s good enough for now.