Support abstracting reified lists
This has a lot of TODOs for more correct error handling, but the
resolution of those should be pretty straightforward once error handling
is more clear overall. The same(value, abstract(reify(value))) tests
work well; I also tried to update the abstract(reified project_name)
test to match the real project_name, where the label is a multilingual
text with a list of monolingual texts, rather than a monolingual text
directly. However, we have to slightly cheat here: the reified nil can’t
be notated in the current JSON, because writing "Z13" will immediately
evaluate to nil, not to a reference with ID "Z13" first, and that
reference is what would be needed for the Z1K2/id pair of the reified
nil. So instead, we get the reified nil from a function call. The rest
of the JSON is all written out, though… excruciatingly long, but with
the variable names, hopefully not completely unreadable.