Port ZObject to dynamic object model
Instead of storing object members in a plain old map, use the dynamic
object model introduced in GraalVM 20.2.0 [1]. It’s not the ideal fit
for us, because our objects are actually pretty static, but it should
already be an improvement over the old code, and make a migration to a
future static object model easier. (Someone in the GraalVM slack
mentioned that a static object model might be coming in future.)

Creation of objects happens through the ZContext, which inherits the
initial shape from the ZLanguage. Not all of the places that used to
call the ZObject constructor had a ZContext already; in some places I
just used lookupContextReference() for now, but I already have some
ideas for how to refactor these nodes so they can use @CachedContext.

ZObject.equals() is completely rewritten, since we can no longer make
use of Map.equals(); on the other hand, now a ZObject can be equal to a
non-ZObject (through interop), which is kind of neat. (Though not very
relevant – at least within GraalEneyj, ZObject.equals() is so far only
called by the Z33/same builtin, which is always chained after the
Z36/value builtin, which already turns interop objects into ZObjects. So
I can’t think of any scenario where ZObject.equals() would actually be
called with a non-ZObject interop object.)

[1]: https://github.com/oracle/graal/blob/master/truffle/docs/DynamicObjectModel.md