Use multiple DynamicObjectLibrary instances in ZPairNode
Use a dedicated instance per static key, as recommended in the
documentation [1]. For a test that calls reify([a, b, c, d, e]) one
million times, this produces a 33% speedup, from ca. 15 seconds to ca.
10 seconds. Nice!

Add a comment to ZObject explaining that this is possible, while
cautioning that objects are conceptually immutable and “external” member
addition should only be done right after the object was created.

[1]: https://github.com/oracle/graal/blob/master/truffle/docs/DynamicObjectModel.md#caching-considerations