Skip intermediate HashMap in ZValueNode
Instead of populating a HashMap with the members only for the ZObject
constructor to add them as dynamic members, add the dynamic members
directly in the node. Because the member names are variable, this
doesn’t seem to make very much of a difference in performance, but I
still think it makes sense to do.