Add experimental WikiLambdaFileSystem
This file system implementation can read from a wiki where the
WikiLambda extension is installed, instead of from disk. I set up such a
wiki at notwikilambda.toolforge.org, and created a Z28 there – though,
for the moment, it’s a plain string and not an object like the eneyj
Z28, because the WikiLambda extension seems to turn any object with type
Z6/string into a plain string on save.