Add another test to WikiLambdaTest
I still haven’t had much success in creating custom objects, but let’s
look at the shape of the generated Z1 for now.

The WikiLambdaTest class will probably be further reorganized in the
future, I’m not very happy with the hard-coded list of paths that go to
Not Wikilambda instead of the real file system.