Migrate from TruffleException to AbstractTruffleException
I have a feeling the error object we delegate to should eventually be a
special kind of object with some extra interop messages (so I guess
ZObjectLiteralNode should look at the type and instantiate a different
class?), but for now a ZObject should do. All tests pass, at least.

Fixes #4.