Add support for Z2/persistent object
A node and runtime class for persistent objects are added (a persistent
object consists of an ID [currently a String, still not completely sure
if it shouldn’t be a ZReference instead], the inner value, and
optionally a labels object). At runtime, a persistent object delegates
all interop messages to the inner value. (This means the label is not
very accessible. I haven’t thought of a good solution to this yet.) A
new helper node is added, ZResolveValueNode, which is similar to
ZEvaluateReferenceNode but also unpacks persistent objects; function
calls and the value builtin use that new node, whereas the reify builtin
continues to use ZEvaluateReferenceNode – I think that reifying a
persistent object should reify the outer object, not the inner value.
(After all, that will then contain the reified inner value, in case you
only need that.) The reify and abstract builtins learn to convert
persistent object runtime values.

WikiLambdaTest::testZTypeMembers() is effectively rewritten – the
members we see through interop are those of the inner type, not those of
the outer persistent object. (The labels currently can’t be accessed,
see above; that the old code described them as “aliases” was always
incorrect.)

We still need the Z1K2/id in a few cases, but should migrate away from
that soon – there isn’t supposed to be any such key anymore.

Note that GraalEneyj is still perfectly happy (as far as I’m aware) to
read stored objects that aren’t persistent objects, though that should
no longer happen.