Display persistent objects using their labels
Add a language/context option for the user interface language, and
expand the display string of persistent objects to look up the label in
that language (if any) and append it to the ID.

I’m not really happy with the code in the ToDisplayString implementation
– I feel like it should live somewhere else, but I’m not sure where
(@ExportMessage classes can’t have @Child nodes, apparently). It’ll do
for now. (It also uses a single uncached InteropLibrary – I tried using
multiple libraries for the different operations but it didn’t seem to
make a difference for performance. I’m not sure if the caching is
working either.)