Parse new implementation layout
Instead of an implementation carrying a Z14K1 member whose type
determines whether it’s a function call, an implementation, or code, in
the Wikifunctions model the implementation directly has a Z14K2, Z14K3
or Z14K4 corresponding to one of those options (though not in this
order). For a function call or code, the object below that key looks
like the Z14K1 value before; for a builtin, the Z14K4 key only serves to
mark the implementation as a builtin – we don’t even look at the Z14K4
value (in real Wikifunctions it would be the ID of the builtin, but our
migrated builtins don’t have IDs), and let the Z14K1 (now meaning the
“implements” of the implementation) tell us what function the builtin
belongs to (this used to be done by the Z1K2 of the builtin object – so
we’ve eliminated another use of Z1K2, yay).