Run code implementations in inner contexts
This ensures that multiple invocations of an implementation won’t share
state, and so our functions will be pure and free of side-effects.

This is implemented by making ZImplementationCodeNode return a new kind
of root node as the call target, instead of the result of parsing the
language source code. The ZInnerContextNode creates an inner context
(from the Env, made available via the ZContext), asks it to parse the
source code (must happen on the inner context – but hopefully the parse
result will be cached and shared across contexts), and then immediately
calls the resulting call target. This must always be an indirect call,
but I don’t see a way around that – Truffle complains if the inner call
target comes from the outer context, so I think we always have to
reparse and therefore always get a new call target. I guess that’s
alright for the time being.

Adds two tests that function implementations can no longer share state
via some global state (globalThis in JavaScript, an arbitrary module in
Python). I tested manually that both tests fail without this commit
(i.e. the JavaScript and Python code is correct).

The context policy of the ZLanguage is changed from “exclusive” to
“shared” – this was necessary because otherwise the ContextReference in
the ZInnerContextNode crashed. As far as I can tell, it’s currently safe
for us to use that context policy (we have no mutable state in the
ZLanguage, and always use @CachedContext in AST nodes), and it’s also
recommended anyways – we’ll just have to watch out in future that we
stay compatible with that context policy (be careful when adding mutable
state to the language instance, don’t directly store ZContext instances
in AST nodes).

I’m not sure whether it’s correct to close the inner context at the end
of the ZInnerContextNode (could it cause issues that we have runtime
objects that outlive the context they were created on?), but it doesn’t
seem to cause any errors so far.