Use dedicated DynamicObjectLibrary for object literal type
The type member is one that we always now we’ll want to set, so it makes
sense to use a dedicated library to optimize that case, I think.

TODO: doesn’t actually seem to help performance…

TODO: but switching from @CachedLibrary to another @Child library like
objectLib seemed to noticeably make performance worse, so maybe try
changing objectLib to a @CachedLibrary too?