Make ZCharacter extend ZObject
This lets us remove the now-redundant has/get language methods, but
requires us to pass a shape into the constructor. Most of the time, we
can get that from the context; in the type system, there doesn’t seem to
be a good source for it (I tried adding a @CachedContext parameter,
didn’t work), but I’m not really convinced that type system is worth it
at all, so I’m okay with putting a very suboptimal solution (new shape
every time) in there for the time being.

ZCharacter still uses a Map for its extra members – I’ll migrate it to
the dynamic object model in a separate commit.