Add option to disable inner contexts
Using inner contexts for code implementations is crucial for
correctness, but also slow. Add an option that allows experts to opt out
of this cost, at their own risk (i.e. we just assume that no
implementations use global state).

The test mentioned in the previous commit message [1] goes from some
three and a half minutes to just over a second with this option
disabled, a speedup of two orders of magnitude.

[1]: https://gist.github.com/lucaswerkmeister/cbb3f682eae730b907368c1f7a4972ee