Introduce runtime class for types
This should make implementing the “meta” interop methods easier in
future (but that isn’t done in this commit yet).

Much of the code is copied and adapted from ZPlainObject /
ZObjectLiteralNode, since for now they function fairly similarly. The
Z4K1/identity is stored in a separate field, though, and we don’t store
the Z1K1/type as a dynamic member either.