Implement meta object methods on ZType
This implements the interop library messages isMetaObject(),
getMetaSimpleName(), getMetaQualifiedName(), and isMetaInstance() on the
ZType runtime type. To implement the latter message, ZObject gains an
abstract method to return the identity of the object’s type as a string.
(I might convert this to a custom Truffle library later, with
@DefaultExports replacing the instanceof check in ZPersistentObject.)

We pretend that Z23/nothing is an instance of Z4/type. According to the
eneyj sources, it should really be an instance of itself, but that would
require it to be a meta object as well, and Truffle doesn’t allow a null
value to also be a meta object. WikiLambda doesn’t have a nothing object
yet, so I hope that the Wikifunctions version of nothing will not be an
instance of itself.

This does not yet implement the getMetaObject() message on regular
objects, but that should not be too hard given getTypeIdentity().

In the test, we build a collection of types and associated values and
assert that values are instances of their associated type and no other
type. (This tests a lot of combinations, but performance doesn’t seem to
suffer too badly.)