Implement getMetaObject() on ZObject
Can be implemented by resolving a reference constructed from the
getTypeIdentity() method introduced in the previous commit. We also
reuse the test set from ZTypeInteropTest, checking that the meta object
returned is the same as the expected type – except for unboxed strings
and booleans, which aren’t ZObjects at all and don’t get a meta object.
I’m not sure if in future we should ensure that unboxed objects don’t
escape beyond a root node, or if it’s just fine and accepted that values
evaluated in our language might not come from our language and might not
have a meta object.