Replace ZObject.getTypeIdentity() with custom library
This lets us get rid of some awkward special cases for unboxed types, by
instead exporting our custom library for them using @DefaultExport.
ZObject has a default implementation of both messages; for the “get”
message, like toDisplayString(), it just throws an exception, and
subclasses are expected to override the message, either by directly
overriding the method with the same syntax, or (as in ZPersistentObject)
by exporting a new version (in this case, with an added cached library).