Support calling Z38/abstract on (reified) types
Reifying types works out of the box thanks to them declaring their
members properly, but turning a reified type back into a type object
requires special support in the abstract builtin. Since the test type,
Z6/string, includes a nested boolean (to indicate whether a key is
required or not), this means we also need to support nested booleans in
reify and abstract. (We don’t support them in the top-level functions
yet, so I’m not adding a test specifically for that.)

And while we’re at it, also update some comments in ZReifyNode.