Implement toDisplayString() on ZType
Forgot to do this when I introduced the class.

Testing it is a bit tricky – to ensure we test the display string of the
type itself, not of the ZPersistentObject wrapping it, we reify it, get
the inner reified persistent object value, abstract it, and then get the
display string of that. (Reifying and abstracting the same value doesn’t
remove a persistent object “layer” on its own.)