Split head/tail into more specializations
This moves the conditionals into guard expressions, which should reduce
the amount of code compiled for call sites that only call with one kind
of list. Parts of the conditions have to be extracted into static helper
methods because guard expressions use a limited language which does not
allow comparison to ZList.NIL or throwing checked exceptions.