Rename JsonElement::getAs*() to as*()
The function names were taken from Gson to simplify the transition away
from that library, but I’ve never liked this aspect – I find it
confusing that getAsJsonObject() and getAsJsonObject(String key) are two
methods with the same name, but one gets the current object whereas the
other gets a member of it. Rename the no-argument versions to as*() to
resolve this ambiguity.