Use @Specialization in ZFunctionCallNode
This is pretty early GraalEneyj code, written when I was less familiar
with Truffle, and in particular didn’t understand specializations as
well yet. This code should be equivalent, but using specialization
parameters is probably cleaner than class fields, and it’ll make it easy
to add other specializations here in the future, if we want to.