Get ZObjects from wikilambda_fetch API instead of action=raw
index.php?action=raw is no longer supported for the ZObject content
model, so we have to use the API instead. We use format=xml instead of
format=json because Java ships built-in XML parsers (several!) but no
built-in JSON parsers; the rudimentary error handling assumes
errorformat=bc because I can’t be bothered to support multiple errors
(the code also assumes that there are no warnings, because otherwise the
first child would be <warnings>).